requirements.txt
boto3==1.34.115
botocore==1.34.115
certifi==2024.2.2
chardet==3.0.4
DateTime==5.5
duckdb==0.10.3
greenlet==3.0.3
idna==2.8
jmespath==1.0.1
mysql-connector-python==8.4.0
numpy==1.26.4
pandas==2.2.2
psycopg2==2.9.9
pyarrow==16.1.0
pymongo==4.7.2
PyMySQL==1.1.1
pyspark==3.5.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.22.0
s3transfer==0.10.1
setuptools==70.0.0
six==1.16.0
sql==2022.4.0
SQLAlchemy==2.0.30
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
xlrd==2.0.1
zope.interface==6.4.post2
